refactor DirImageSource to return just an image
This changed the ImageSource in such was that instead of
returning a tuple with the image and a string (label),
only an image is returned.

The subdirectory, in which the image was found, as well as
the image name without file extension is instead added as
metadata (property) to the image itself.