refactor ImageOperation to use ProbableOperation
ProbableOperation effectively separates the functionality
of having an operation only applied sometimes (depending
on chance) from the individual operations.
That means that operations such as FlipX are now deterministic.

That being said, those operations now have a custom constructor
that lifts them into a ProbableOperation if a chance is specified

This patch lays the groundwork for transforming two images
the same way, even if ProbableOperations are involved.

Tests included.